FileRename anr refactor
